Add view model file.
